Implement hybrid precision for GTX GPUs
- Use FP16 for RAFT flow computation (speed)
- Keep inpainting model in FP32 (quality - prevents black spots)
- Best of both: faster processing + clean watermark removal
- Fixes GTX 1650/1660 black spot issue without full FP32 slowdown

+++ faster-propainter-main/watermark.py
@@ -360,7 +360,8 @@ def pipeline(
             )
             gt_flows_bi = (gt_flows_bi[0].half(), gt_flows_bi[1].half())
             fix_flow_complete = fix_flow_complete.half()
-            model = model.half()
+            # SKIP model.half() - keep inpainting model in FP32 for quality on GTX GPUs
+            # model = model.half()
 
         # ---- complete flow ----
         flow_length = gt_flows_bi[0].size(1)
